make ghResetHard even safer

+++ GitHelp/ghResetHard.sh
@@ -13,4 +13,12 @@ then
     exit 1
 fi
 
+read -p "Are you REALLY sure?  (y/n)   " -n 1 -r
+echo
+if [[ ! $REPLY =~ ^[Yy]$ ]]
+then
+    printf "\nOperation canceled.\n\n"
+    exit 1
+fi
+
 git reset --hard
